Split MTransaction and download classes
